+++ API.Tests/Utilities/OtrClaimsTests.cs
@@ -1,3 +1,4 @@
+using API.Authorization;
 using API.Utilities;
 
 namespace APITests.Utilities;
