+++ API.Tests/Utilities/RateLimitOverridesSerializerTests.cs
@@ -1,3 +1,4 @@
+using API.Authorization;
 using API.Utilities;
 using Database.Entities;
