+++ API/Controllers/MatchesController.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel.DataAnnotations;
+using API.Authorization;
 using API.DTOs;
 using API.Services.Interfaces;
 using API.Utilities;
